+++ astrometrics/urls.py
@@ -12,6 +12,8 @@ urlpatterns = [
     path("manager/", include("manager.urls")),
     path("manager/cms/", include("cms.manager_urls")),
     path("manager/catalogue/", include("journals.manager_urls")),
+    path("manager/portal/", include("portal.manager_urls")),
+    path("portal/", include("portal.urls")),
     path("catalogue/", include("journals.urls")),
     path("", include("cms.urls")),
 ]